Add testing and documentation to FixedMath
There are no logical or execution changes to FixedMath in this change.
This change adds documentation to Fixed math about it’s fixed point
system, brings the formatting to a similar standard of the rest of the
codebase, and adds Math/Tests/Editor/FixedMathTest.cs a unit test class
that tests the majority of the features of FixedMath and can be run in
Unity by browsing to window->EditorTestsRunner. The tests should
automatically load and be able to be run in the project.